third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-preserveAlpha-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'preserveAlpha' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.preserveAlpha.baseVal is false
PASS convolveMatrixElement.preserveAlpha.baseVal is true
PASS successfullyParsed is true

TEST COMPLETE
